mv perf_patient fixture to conftest; add client to init fixtures
